+++ packages/react-ui/src/index.d.ts
@@ -39,7 +39,6 @@ import Typography, {
 import Button, { ButtonProps } from './components/atoms/Button';
 import PasswordField, { PasswordFieldProps } from './components/atoms/PasswordField';
 import SelectField, { SelectFieldProps } from './components/atoms/SelectField';
-import SelectField2, { SelectFieldProps2 } from './components/atoms/SelectField2';
 import MultipleSelectField, {
   MultipleSelectFieldProps
 } from './components/atoms/MultipleSelectField';
@@ -106,8 +105,6 @@ export {
   PasswordFieldProps,
   SelectField,
   SelectFieldProps,
-  SelectField2,
-  SelectFieldProps2,
   MultipleSelectField,
   MultipleSelectFieldProps,
   UploadField,
