+++ packages/react-ui/src/types.d.ts
@@ -1,6 +1,5 @@
 import { GroupDateTypes } from '@carto/react-core';
 export { SelectFieldProps } from './components/atoms/SelectField';
-export { SelectFieldProps2 } from './components/atoms/SelectField2';
 export { MultipleSelectFieldProps } from './components/atoms/MultipleSelectField';
 export { TypographyProps } from './components/atoms/Typography';
 export { LabelWithIndicatorProps } from './components/atoms/LabelWithIndicator';
