+++ packages/react-ui/storybook/stories/atoms/SelectField.stories.js
@@ -10,7 +10,6 @@ import {
   Label
 } from '../../utils/storyStyles';
 import Button from '../../../src/components/atoms/Button';
-import { EuroOutlined } from '@mui/icons-material';
 
 const options = {
   title: 'Atoms/Select Field',
@@ -70,18 +69,9 @@ const options = {
 export default options;
 
 const menuItems = [
-  {
-    label: 'Ten: super large text with overflow',
-    value: '10'
-  },
-  {
-    label: 'Twenty',
-    value: '20'
-  },
-  {
-    label: 'Thirty',
-    value: '30'
-  }
+  { label: 'Ten: super large text with overflow', id: '10' },
+  { label: 'Twenty', id: '20' },
+  { label: 'Thirty', id: '30' }
 ];
 
 const SelectFieldItem = ({
@@ -108,80 +98,27 @@ const SelectFieldItem = ({
     );
   };
 
-  const optionsSelect = [
-    { label: 'Ten: super large text with overflow', value: '10', icon: <EuroOutlined /> },
-    { label: 'Twenty', value: '20', icon: <EuroOutlined /> },
-    { label: 'Thirty', value: '30', icon: <EuroOutlined /> }
-  ];
-
   return (
-    <div>
-      <div>
-        <Typography variant='body1' mt={4}>
-          {'Object'}
-        </Typography>
-        <SelectField
-          {...rest}
-          label={label}
-          variant={variant}
-          placeholder={placeholder}
-          helperText={helperText}
-          value={content}
-          focused={focused}
-          disabled={disabled}
-          onChange={handleChange}
-          error={error}
-          size={size}
-          fullWidth={rest.fullWidth}
-          defaultValue={optionsSelect.length && optionsSelect[0].label}
-        >
-          {optionsSelect.map((o, i) => (
-            <MenuItem key={i} value={typeof o === 'object' ? o.value : o}>
-              {typeof o === 'object' ? (
-                <span>
-                  {o.icon && <img src={o.icon} alt='' />}
-                  {o.label}
-                </span>
-              ) : (
-                o
-              )}
-            </MenuItem>
-          ))}
-        </SelectField>
-      </div>
-      <div>
-        <Typography variant='body1' mt={4}>
-          {'Array'}
-        </Typography>
-        <SelectField
-          {...rest}
-          label={label}
-          variant={variant}
-          placeholder={placeholder}
-          helperText={helperText}
-          onChange={handleChange}
-          value={content}
-          focused={focused}
-          disabled={disabled}
-          error={error}
-          size={size}
-          fullWidth={rest.fullWidth}
-        >
-          {[...Array(20)].map((item, index) => {
-            const itemText =
-              index === 1
-                ? `Very long item text with overflow ${index + 1}`
-                : `Item ${index + 1}`;
-
-            return (
-              <MenuItem key={index} value={itemText}>
-                {itemText}
-              </MenuItem>
-            );
-          })}
-        </SelectField>
-      </div>
-    </div>
+    <SelectField
+      {...rest}
+      label={label}
+      variant={variant}
+      placeholder={placeholder}
+      helperText={helperText}
+      value={content}
+      focused={focused}
+      disabled={disabled}
+      onChange={handleChange}
+      error={error}
+      size={size}
+      fullWidth={rest.fullWidth}
+    >
+      {menuItems.map((o, i) => (
+        <MenuItem key={i} value={o.label}>
+          {o.label}
+        </MenuItem>
+      ))}
+    </SelectField>
   );
 };
